+++ src/library/scala/xml/dtd/DocType.scala
@@ -18,8 +18,7 @@ package dtd
  *  @param  extID  NoExternalID or the external ID of this doctype
  *  @param  intSubset sequence of internal subset declarations
  */
-case class DocType(name: String, extID: ExternalID, intSubset: Seq[dtd.Decl])
-{
+case class DocType(name: String, extID: ExternalID, intSubset: Seq[dtd.Decl]) {
   if (!Utility.isName(name))
     throw new IllegalArgumentException(name+" must be an XML Name")
 
@@ -33,8 +32,7 @@ case class DocType(name: String, extID: ExternalID, intSubset: Seq[dtd.Decl])
   }
 }
 
-object DocType
-{
+object DocType {
   /** Creates a doctype with no external id, nor internal subset declarations. */
   def apply(name: String): DocType = apply(name, NoExternalID, Nil)
 }
